Fix handling of blank commands
When a potential system pager passed in via
the environment variables or git configuration
yields a pager that is blank ("") or whitespace-
only ("  ") the call to `TTY::Which.exist?`
would raise an error.

This behavior could be fixed in either gem,
however I was wary of having `TTY::Which`
gracefully handle this case and then have
dependencies (including `TTY::Pager`) need
to know about its behavior when `nil` is
passed; as a result fixing `TTY::Pager`
seemed cleaner to me.

Closes #5.