Change to extract abstract pager and convert Pager to a module #16
This extracts a common TTY::Pager::Abstract interface that requires overriding
write, puts and close methods. Changed Pager to be a module since it no longer
requires instatiation. Kept the #new method as a factory for the null, basic
and system pager instances. Made the #select_pager private to protect against
future changes. Only the top-level #page method remains as a convenience for
instance paging.