Extract abstract pager and convert Pager to a module #16 (#20)
Extract a common TTY::Pager::Abstract interface that requires overriding
write, puts and close methods. Change Pager to be a module as it no longer
needs instantiation. Keep the #new method as a factory for the null, basic
and system pager instances. Make the #select_pager private to protect against
future changes. Only the top-level #page method remains as a convenience for
instance paging.